Fix broken pipe errors and timeout handling (Issue #30)

+++ src/protocol/constants.rs
@@ -1,6 +1,6 @@
 pub const PROTOCOL_VERSION: u8 = 7;
-pub const ACTIVITY_TIMEOUT: u32 = 120;
-pub const PONG_TIMEOUT: u32 = 30;
+pub const ACTIVITY_TIMEOUT: u64 = 120;
+pub const PONG_TIMEOUT: u64 = 30;
 
 pub const CHANNEL_NAME_MAX_LENGTH: usize = 200;
 pub const CHANNEL_NAME_REGEX: &str = r"^[a-zA-Z0-9_\-=@,.;]+$";
